Refactor evidence overview: implement loading state, replace useEvidenceDashboard hook with direct data fetching, and enhance error handling. Introduce new data structure for evidence dashboard and remove obsolete action file.
