+++ apps/app/src/app/[locale]/(app)/(dashboard)/evidence/overview/hooks/useEvidenceData.ts
@@ -1,8 +1,8 @@
 import { useI18n } from "@/locales/client";
 import type { Departments } from "@prisma/client";
-import type { EvidenceWithStatus } from "../actions/getEvidenceDashboard";
 import { DEPARTMENT_NAMES } from "../constants/departments";
 import type { StatusType } from "../constants/evidence-status";
+import type { EvidenceWithStatus } from "../data/getEvidenceDashboard";
 
 /**
  * Interface for the processed department data
