+++ apps/app/src/app/[locale]/(app)/(dashboard)/evidence/overview/hooks/useFrameworkData.ts
@@ -1,4 +1,4 @@
-import type { EvidenceWithStatus } from "../actions/getEvidenceDashboard";
+import type { EvidenceWithStatus } from "../data/getEvidenceDashboard";
 import type { StatusType } from "../constants/evidence-status";
 
 /**
