+++ apps/app/src/app/[locale]/(app)/(dashboard)/evidence/overview/loading.tsx
@@ -1,0 +1,5 @@
+import { EvidenceOverviewSkeleton } from "./components/EvidenceUIStates";
+
+export default function Loading() {
+	return <EvidenceOverviewSkeleton />;
+}
