+++ apps/app/src/lib/errors.ts
@@ -1,0 +1,17 @@
+export class AppError extends Error {
+  constructor(
+    public code: string,
+    message?: string,
+    public status: number = 400
+  ) {
+    super(message || code);
+    this.name = "AppError";
+  }
+}
+
+export const appErrors = {
+  UNAUTHORIZED: "UNAUTHORIZED",
+  NOT_FOUND: "NOT_FOUND",
+  VALIDATION_ERROR: "VALIDATION_ERROR",
+  UNEXPECTED_ERROR: "UNEXPECTED_ERROR",
+} as const;
